server/python/requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
pillow>=8.0.0
pytesseract>=0.3.8
ultralytics>=8.0.0
